move dbtypes to db dir
