+++ backend/db/problems/getProblemFromId.go
@@ -2,7 +2,6 @@ package problems
 
 import (
 	"context"
-	"go-test/myTypes"
 	"log"
 	"net/http"
 	"os"
@@ -12,11 +11,11 @@ import (
 	"go.mongodb.org/mongo-driver/mongo"
 )
 
-func GetProblemFromId(c *gin.Context, problemId int32) myTypes.Problem {
+func GetProblemFromId(c *gin.Context, problemId int32) Problem {
 	client, exists := c.Get("mongoClient")
 	if !exists {
 		c.JSON(http.StatusInternalServerError, gin.H{"Error": "DB client is not available."})
-		return myTypes.Problem{}
+		return Problem{}
 	}
 	dbName := os.Getenv("DB_NAME")
 	prbCol := os.Getenv("PROBLEMS_COLLECTION")
@@ -24,11 +23,11 @@ func GetProblemFromId(c *gin.Context, problemId int32) myTypes.Problem {
 
 	filter := bson.M{"problemId": problemId}
 
-	var problem myTypes.Problem
+	var problem Problem
 	err := collection.FindOne(context.TODO(), filter).Decode(&problem)
 	if err != nil {
 		log.Printf("Failed to find single result from DB: %v\n", err.Error())
-		return myTypes.Problem{}
+		return Problem{}
 	}
 	return problem
 }
